chore: separate Map and Export
