+++ src/utilities/codeEditorUtils.js
@@ -6,6 +6,8 @@ const path = require('path');
 const os = require('os');
 const fs = require('fs');
 const { spawnSync } = require('child_process');
+import { EEMap } from "./eeMap";
+import { EEExport } from "./eeExport";
 import { EEChart } from "./eeChart";
 import { EEUi } from "./eeUi";
 
@@ -49,247 +51,15 @@ function tryGitClone(eeRepository, repoPath){
     } 
 }
 
-/* ExportImage: wrapper for ee.batchExport.image.toXXX 
-functions, but also starts the tasks automatically.
-*/
-class ExportImage {
-    constructor(ee, successCallback, errCallback){
-        this.toAsset = function(...args){
-          var computed = false;
-          var clientConfig = ee.arguments.extractFromFunction(
-              ee.batch.Export.image.toAsset, arguments);
-          if(!Object.hasOwn(clientConfig,"description")){
-              clientConfig["description"] = "myEETasksExportImageTask";
-          }
-          //🔲 TODO: assetId default to
-          // projects/PROJECT/assets/ + description
-          if (Object.hasOwn(clientConfig, "region")){
-            var region = clientConfig["region"];
-            if (region.func){
-              computed = true;
-              region.evaluate(
-                  (r)=>{
-              clientConfig["region"] = r;
-              return ee.batch.Export.image.toAsset(clientConfig)
-              .start(successCallback, errCallback);
-                  }
-              );
-            }
-            if (!computed){
-            return ee.batch.Export.image.toAsset(clientConfig)
-            .start(successCallback, errCallback);
-           }
-           }
-        };
-
-        this.toCloudStorage = function(...args){
-          var computed = false;
-          var clientConfig = ee.arguments.extractFromFunction(
-              ee.batch.Export.image.toCloudStorage, arguments);
-          if(!Object.hasOwn(clientConfig,"description")){
-              clientConfig["description"] = "myEETasksExportImageTask";
-          }
-          if(!Object.hasOwn(clientConfig,"fileNamePrefix")){
-              clientConfig["fileNamePrefix"] = clientConfig["description"];
-          }
-          if (Object.hasOwn(clientConfig, "region")){
-            var region = clientConfig["region"];
-            if (region.func){
-              computed=true;
-              region.evaluate(
-                  (r)=>{
-              clientConfig["region"] = r;
-              return ee.batch.Export.image.toCloudStorage(clientConfig)
-              .start(successCallback, errCallback);
-                  }
-              );
-            }
-          }
-          if (!computed){
-            return ee.batch.Export.image.toCloudStorage(clientConfig)
-            .start(successCallback, errCallback);
-          }
-        };
-
-        this.toDrive = function(...args){
-          var computed = false;
-          var clientConfig = ee.arguments.extractFromFunction(
-              ee.batch.Export.image.toDrive, arguments);
-          if(!Object.hasOwn(clientConfig,"description")){
-              clientConfig["description"] = "myEETasksExportImageTask";
-          }
-          if(!Object.hasOwn(clientConfig,"fileNamePrefix")){
-              clientConfig["fileNamePrefix"] = clientConfig["description"];
-          }
-          if (Object.hasOwn(clientConfig, "region")){
-            var region = clientConfig["region"];
-            if (region.func){
-              computed=true;
-              region.evaluate(
-                  (r)=>{
-              clientConfig["region"] = r;
-              return ee.batch.Export.image.toDrive(clientConfig)
-              .start(successCallback, errCallback);
-                  }
-              );
-            }
-          }
-          if (!computed){
-            return ee.batch.Export.image.toDrive(clientConfig)
-            .start(successCallback, errCallback);
-          }
-        };
-    }
-}
-
-/* ExportMap: wrapper for ee.batchExport.map.toXXX 
-functions, but also starts the tasks automatically.
-🔲 TODO: defaults?
-*/
-class ExportMap {
-    constructor(ee, successCallback, errCallback){
-        this.toCloudStorage = function(...args){
-            return ee.batch.Export.map.toCloudStorage(...args)
-            .start(successCallback, errCallback);
-        };
-    }
-}
-
-/* ExportVideo: wrapper for ee.batchExport.video.toXXX 
-functions, but also starts the tasks automatically.
-🔲 TODO: defaults?
-*/
-class ExportVideo {
-    constructor(ee, successCallback, errCallback){
-        this.toCloudStorage = function(...args){
-            return ee.batch.Export.video.toCloudStorage(...args)
-            .start(successCallback, errCallback);
-        };
-    }
-}
-
-/* ExportTable: wrapper for ee.batchExport.table.toXXX 
-functions, but also starts the tasks automatically.
-🔲 TODO: description default to myExportTableTask
-*/
-class ExportTable {
-    constructor(ee, successCallback, errCallback){
-        this.toAsset = function(...args){
-        //🔲 TODO: assetId default to
-        // projects/PROJECT/assets/ + description
-            return ee.batch.Export.table.toAsset(...args)
-            .start(successCallback, errCallback);
-        };
-        this.toCloudStorage = function(...args){
-        //🔲 TODO: fileNamePrefix default to description
-        //🔲 TODO: bucket default?
-            return ee.batch.Export.table.toCloudStorage(...args)
-            .start(successCallback, errCallback);
-        };
-        this.toDrive = function(...args){
-        //🔲 TODO: fileNamePrefix default to description
-            return ee.batch.Export.table.toDrive(...args)
-            .start(successCallback, errCallback);
-        };
-        this.toBigQuery = function(...args){
-        //🔲 TODO: defaults?
-            return ee.batch.Export.table.toBigQuery(...args)
-            .start(successCallback, errCallback);
-        };
-        this.toFeatureView = function(...args){
-        //🔲 TODO: defaults?
-            return ee.batch.Export.table.toFeatureView(...args)
-            .start(successCallback, errCallback);
-        };
-    }
-}
-
-class ExportConstructor{
-    constructor(ee, successCallback, errCallback){
-        this.table = new ExportTable(ee, successCallback, errCallback);
-        this.image = new ExportImage(ee, successCallback, errCallback);
-        this.map = new ExportMap(ee, successCallback, errCallback);
-        this.video = new ExportVideo(ee, successCallback, errCallback);
-    }
-}
-
- class MapConstructor{
-    constructor(ee, panel, uri){
-        this.ee = ee;
-        this._panel = undefined;
-        this._openMapPanelIfNeeded = function(){
-        if (typeof this._vsMapPanel === 'undefined'){
-            this._vsMapPanel = panel.render(uri);
-        }
-        };
-        this.setCenter=function(lon,lat,zoom){
-            this._openMapPanelIfNeeded();
-            const coord = [lat, lon];
-            this._vsMapPanel.setView(coord, zoom);
-        };
-        this.addLayer=function(eeObject,visParams,name,shown,opacity){
-            this._openMapPanelIfNeeded();
-            let ee = this.ee;
-            if (typeof eeObject.mosaic === 'function'){
-                // ImageCollection has a `mosaic` method:
-                eeObject = eeObject.mosaic(); 
-            }else{
-                // Both geometry and feature have a `centroid` method
-                // FeatureCollection has the `aggregate_array` method
-                if(typeof eeObject.centroid === 'function' || typeof eeObject.aggregate_array === 'function'){
-                    var features = ee.FeatureCollection(eeObject);
-                    var color = visParams && visParams.color ? visParams.color : '000000';
-                    var width = visParams && visParams.width ? visParams.width : 2;
-                    var imageOutline = features.style({
-                        color: color,
-                        fillColor: '00000000',
-                        width: width
-                    });
-                    eeObject = features.style({fillColor: color})
-                    .updateMask(ee.Image.constant(0.5))
-                    .blend(imageOutline);
-                }
-            }
-            const request = ee.data.images.applyVisualization(eeObject, visParams);
-            const mapId = ee.data.getMapId(request);
-            this._vsMapPanel.addLayer(mapId.urlFormat, name, shown, opacity);
-        };   
-        // NOT YET IMPLEMENTED:
-        this.add=function(item){};
-        this.centerObject=function(object,zoom,onComplete){};
-        this.clear=function(){};
-        this.drawingTools=function(){};
-        this.getBounds=function(asGeoJSON){};
-        this.getCenter=function(){};
-        this.getScale=function(){};
-        this.getZoom=function(){};
-        this.layers=function(){};
-        this.onChangeBounds=function(callback){};
-        this.onChangeCenter=function(callback){};
-        this.onChangeZoom=function(callback){};
-        this.onClick=function(callback){};
-        this.onIdle=function(callback){};
-        this.onTileLoaded=function(callback){};
-        this.remove=function(item){};
-        this.setControlVisibility=function(all,layerList,zoomControl,scaleControl,
-         mapTypeControl,fullscreenControl,drawingToolsControl){};
-        this.setGestureHandling=function(option){};
-        this.setZoom=function(zoom){};
-        this.style=function(){};
-        this.unlisten=function(idOrType){};
-        this.widgets=function(){};
-    }
-};
-
 class Tools {
-    constructor(ee, vsLog, vsMap, vsUri, successCallback, errCallback, fileName) {
+    constructor(ee, vsLog, vsUri, successCallback, errCallback, fileName) {
         this.fileName = fileName;
         this.filePath = path.dirname(fileName);
         this.ee = ee;
         this.log = this.Log(vsLog);
         this.print = this.Print(this.log);
-        this.Map = new MapConstructor(ee, vsMap, vsUri);
-        this.Export = new ExportConstructor(ee, successCallback, errCallback);
+        this.Map = new EEMap(ee, vsUri);
+        this.Export = new EEExport(ee, successCallback, errCallback);
 
         /* These are defined to be silently ignored*/
         this.ui = EEUi;
