+++ src/utilities/eeExport.ts
@@ -1,0 +1,178 @@
+/* ExportImage: wrapper for ee.batchExport.image.toXXX 
+functions, but also starts the tasks automatically.
+*/
+class ExportImage {
+    toAsset: (...args: any[]) => any;
+    toCloudStorage: (...args: any[]) => any;
+    toDrive: (...args: any[]) => any;
+    constructor(ee: any, successCallback: any, errCallback: any){
+        this.toAsset = function(...args){
+          var computed = false;
+          var clientConfig = ee.arguments.extractFromFunction(
+              ee.batch.Export.image.toAsset, arguments);
+          if(!Object.hasOwn(clientConfig,"description")){
+              clientConfig["description"] = "myEETasksExportImageTask";
+          }
+          //🔲 TODO: assetId default to
+          // projects/PROJECT/assets/ + description
+          if (Object.hasOwn(clientConfig, "region")){
+            var region = clientConfig["region"];
+            if (region.func){
+              computed = true;
+              region.evaluate(
+                  (r: any)=>{
+              clientConfig["region"] = r;
+              return ee.batch.Export.image.toAsset(clientConfig)
+              .start(successCallback, errCallback);
+                  }
+              );
+            }
+            if (!computed){
+            return ee.batch.Export.image.toAsset(clientConfig)
+            .start(successCallback, errCallback);
+           }
+           }
+        };
+
+        this.toCloudStorage = function(...args){
+          var computed = false;
+          var clientConfig = ee.arguments.extractFromFunction(
+              ee.batch.Export.image.toCloudStorage, arguments);
+          if(!Object.hasOwn(clientConfig,"description")){
+              clientConfig["description"] = "myEETasksExportImageTask";
+          }
+          if(!Object.hasOwn(clientConfig,"fileNamePrefix")){
+              clientConfig["fileNamePrefix"] = clientConfig["description"];
+          }
+          if (Object.hasOwn(clientConfig, "region")){
+            var region = clientConfig["region"];
+            if (region.func){
+              computed=true;
+              region.evaluate(
+                  (r: any)=>{
+              clientConfig["region"] = r;
+              return ee.batch.Export.image.toCloudStorage(clientConfig)
+              .start(successCallback, errCallback);
+                  }
+              );
+            }
+          }
+          if (!computed){
+            return ee.batch.Export.image.toCloudStorage(clientConfig)
+            .start(successCallback, errCallback);
+          }
+        };
+
+        this.toDrive = function(...args){
+          var computed = false;
+          var clientConfig = ee.arguments.extractFromFunction(
+              ee.batch.Export.image.toDrive, arguments);
+          if(!Object.hasOwn(clientConfig,"description")){
+              clientConfig["description"] = "myEETasksExportImageTask";
+          }
+          if(!Object.hasOwn(clientConfig,"fileNamePrefix")){
+              clientConfig["fileNamePrefix"] = clientConfig["description"];
+          }
+          if (Object.hasOwn(clientConfig, "region")){
+            var region = clientConfig["region"];
+            if (region.func){
+              computed=true;
+              region.evaluate(
+                  (r: any)=>{
+              clientConfig["region"] = r;
+              return ee.batch.Export.image.toDrive(clientConfig)
+              .start(successCallback, errCallback);
+                  }
+              );
+            }
+          }
+          if (!computed){
+            return ee.batch.Export.image.toDrive(clientConfig)
+            .start(successCallback, errCallback);
+          }
+        };
+    }
+}
+
+/* ExportTable: wrapper for ee.batchExport.table.toXXX 
+functions, but also starts the tasks automatically.
+🔲 TODO: description default to myExportTableTask
+*/
+class ExportTable {
+    toAsset: (...args: any[]) => any;
+    toCloudStorage: (...args: any[]) => any;
+    toDrive: (...args: any[]) => any;
+    toBigQuery: (...args: any[]) => any;
+    toFeatureView: (...args: any[]) => any;
+    constructor(ee:any , successCallback:any , errCallback:any ){
+        this.toAsset = function(...args){
+        //🔲 TODO: assetId default to
+        // projects/PROJECT/assets/ + description
+            return ee.batch.Export.table.toAsset(...args)
+            .start(successCallback, errCallback);
+        };
+        this.toCloudStorage = function(...args){
+        //🔲 TODO: fileNamePrefix default to description
+        //🔲 TODO: bucket default?
+            return ee.batch.Export.table.toCloudStorage(...args)
+            .start(successCallback, errCallback);
+        };
+        this.toDrive = function(...args){
+        //🔲 TODO: fileNamePrefix default to description
+            return ee.batch.Export.table.toDrive(...args)
+            .start(successCallback, errCallback);
+        };
+        this.toBigQuery = function(...args){
+        //🔲 TODO: defaults?
+            return ee.batch.Export.table.toBigQuery(...args)
+            .start(successCallback, errCallback);
+        };
+        this.toFeatureView = function(...args){
+        //🔲 TODO: defaults?
+            return ee.batch.Export.table.toFeatureView(...args)
+            .start(successCallback, errCallback);
+        };
+    }
+}
+
+
+/* ExportMap: wrapper for ee.batchExport.map.toXXX 
+functions, but also starts the tasks automatically.
+🔲 TODO: defaults?
+*/
+class ExportMap {
+    toCloudStorage: (...args: any[]) => any;
+    constructor(ee:any, successCallback:any , errCallback:any){
+        this.toCloudStorage = function(...args){
+            return ee.batch.Export.map.toCloudStorage(...args)
+            .start(successCallback, errCallback);
+        };
+    }
+}
+
+/* ExportVideo: wrapper for ee.batchExport.video.toXXX 
+functions, but also starts the tasks automatically.
+🔲 TODO: defaults?
+*/
+class ExportVideo {
+    toCloudStorage: (...args: any[]) => any;
+    constructor(ee: any, successCallback:any , errCallback:any ){
+        this.toCloudStorage = function(...args){
+            return ee.batch.Export.video.toCloudStorage(...args)
+            .start(successCallback, errCallback);
+        };
+    }
+}
+
+export class EEExport{
+    table: ExportTable;
+    image: ExportImage;
+    map: ExportMap;
+    video: ExportVideo;
+    constructor(ee:any , successCallback:any , errCallback:any){
+        this.table = new ExportTable(ee, successCallback, errCallback);
+        this.image = new ExportImage(ee, successCallback, errCallback);
+        this.map = new ExportMap(ee, successCallback, errCallback);
+        this.video = new ExportVideo(ee, successCallback, errCallback);
+    }
+}
