+++ tsconfig.json
@@ -1,9 +1,9 @@
 {
 	"compilerOptions": {
 		"module": "commonjs",
-		"target": "ES2020",
+		"target": "ES2022",
 		"outDir": "out",
-		"lib": ["ES2020", "DOM"],
+		"lib": ["ES2022", "DOM"],
 		"sourceMap": true,
 		"rootDir": "src",
 		"strict": true   /* enable all strict type-checking options */
